Finished to write a test for the parsing of the input file for GheoNet engine.

+++ it.unifi.dsi.stlab.networkreasoner.gas.system.formulae/variants/KvalueFormula.cs
@@ -34,7 +34,7 @@ namespace it.unifi.dsi.stlab.networkreasoner.gas.system.formulae
 			set;
 		}
 
-		public long EdgeLength {
+		public double EdgeLength {
 			get;
 			set;
 		}
