+++ it.unifi.dsi.stlab.networkreasoner.gas.system/exactly-dimensioned-instance/EdgeForNetwonRaphsonSystem.cs
@@ -283,7 +283,7 @@ namespace it.unifi.dsi.stlab.networkreasoner.gas.system.exactly_dimensioned_inst
 
 		public EdgeState SwitchState{ get; set; }
 
-		public long Length { get; set; }
+		public double Length { get; set; }
 
 		public double DiameterInMillimeters { get; set; }
